Rename, rearrange and clarify exports of serialisation
